+++ docs/app/_components/SiteHeader.tsx
@@ -5,7 +5,7 @@ import { SiteMenu } from './SiteMenu';
 import { SiteNavigation } from './SiteNavigation';
 
 export const SiteHeader = () => (
-  <header className="border-secondary-200 bg-bg-body sticky top-0 z-10 flex h-(--page-header-height) w-full items-center gap-2 border-b px-(--page-padding) md:px-(--page-padding-md) xl:px-(--page-padding-xl)">
+  <header className="border-secondary-200 bg-bg-body sticky top-0 z-50 flex h-(--page-header-height) w-full items-center gap-2 border-b px-(--page-padding) md:px-(--page-padding-md) xl:px-(--page-padding-xl)">
     <div className="md:hidden">
       <MobileNavigation />
     </div>