+++ docs/content/patterns/forms/form-action.demo.tsx
@@ -1,7 +1,7 @@
 import { Button, Inline } from '@marigold/components';
 
 export default () => (
-  <Inline space="fieldX" alignY="center">
+  <Inline space="related" alignY="center">
     <Button variant="primary">Save</Button>
     <Button variant="secondary">Cancel</Button>
   </Inline>