+++ packages/components/src/Card/Card.stories.tsx
@@ -101,18 +101,21 @@ export const Stretch: Story = {
   },
   render: args => (
     <Card {...args}>
-      <Headline level="2">Professor Severus Snape</Headline>
-      <Text>
-        Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam nonumy
-        eirmod tempor invidunt ut labore et dolore magna aliquyam erat, sed diam
-        voluptua. At vero eos et accusam et justo duo dolores et ea rebum. Stet
-        clita kasd gubergren, no sea takimata sanctus est Lorem ipsum dolor sit
-        amet. Lorem ipsum dolor sit amet, consetetur sadipscing elitr, sed diam
-        nonumy eirmod tempor invidunt ut labore et dolore magna aliquyam erat,
-        sed diam voluptua. At vero eos et accusam et justo duo dolores et ea
-        rebum. Stet clita kasd gubergren, no sea takimata sanctus est Lorem
-        ipsum dolor sit amet.
-      </Text>
+      <Container>
+        <Headline level="2">Professor Severus Snape</Headline>
+      </Container>
+      <Container contentLength="long">
+        <Text>
+          <strong>Professor Severus Snape</strong> (9 January, 1960[1] - 2 May,
+          1998)[2] was an English half-blood[3] wizard serving as Potions Master
+          (1981-1996), Head of Slytherin House (1981-1997), Defence Against the
+          Dark Arts professor (1996-1997), and Headmaster (1997-1998) of the
+          Hogwarts School of Witchcraft and Wizardry as well as a member of the
+          Order of the Phoenix and a Death Eater. His double life played an
+          extremely important role in both of the Wizarding Wars against
+          Voldemort.
+        </Text>
+      </Container>
     </Card>
   ),
 };